Fix signed/unsigned mismatch warning.

+++ src/Core/OgreRenderingModule/RenderWindow.cpp
@@ -185,7 +185,7 @@ void RenderWindow::Resize(int width, int height)
 {
     assert(renderWindow);
 
-    if (width == renderWindow->getWidth() && height == renderWindow->getHeight())
+    if (width == (int)renderWindow->getWidth() && height == (int)renderWindow->getHeight())
         return; // Avoid recreating resources if the size didn't actually change.
 
     renderWindow->resize(width, height);
